+++ ElementX/Sources/Services/Notification/Proxy/NotificationItemProxy.swift
@@ -14,92 +14,9 @@
 // limitations under the License.
 //
 
-import CryptoKit
 import Foundation
-import UserNotifications
-
 import MatrixRustSDK
-
-protocol NotificationItemProxyProtocol {
-    var event: NotificationEvent? { get }
-
-    var eventID: String { get }
-
-    var senderID: String { get }
-
-    var roomID: String { get }
-
-    var receiverID: String { get }
-
-    var senderDisplayName: String? { get }
-
-    var senderAvatarMediaSource: MediaSourceProxy? { get }
-
-    var roomDisplayName: String { get }
-
-    var roomCanonicalAlias: String? { get }
-
-    var roomAvatarMediaSource: MediaSourceProxy? { get }
-
-    var roomJoinedMembers: Int { get }
-
-    var isRoomDirect: Bool { get }
-
-    var isNoisy: Bool { get }
-}
-
-extension NotificationItemProxyProtocol {
-    var isEncrypted: Bool {
-        switch event {
-        case .none, .invite:
-            return false
-        case .timeline(let event):
-            switch try? event.eventType() {
-            case .messageLike(let content):
-                switch content {
-                case .roomEncrypted:
-                    return true
-                default:
-                    return false
-                }
-            default:
-                return false
-            }
-        }
-    }
-
-    var isDM: Bool {
-        isRoomDirect && roomJoinedMembers <= 2
-    }
-    
-    var hasMedia: Bool {
-        if (isDM && senderAvatarMediaSource != nil) ||
-            (!isDM && roomAvatarMediaSource != nil) {
-            return true
-        }
-        switch event {
-        case .invite, .none:
-            return false
-        case .timeline(let event):
-            switch try? event.eventType() {
-            case .state, .none:
-                return false
-            case let .messageLike(content):
-                switch content {
-                case let .roomMessage(messageType):
-                    switch messageType {
-                    case .image, .video, .audio:
-                        return true
-                    default:
-                        return false
-                    }
-                default:
-                    return false
-                }
-            }
-        }
-    }
-}
+import UserNotifications
 
 struct NotificationItemProxy: NotificationItemProxyProtocol {
     let notificationItem: NotificationItem
